task2/requirements.txt
langchain 
streamlit 
faiss-cpu 
sentence-transformers 
ollama
langchain-community